Added support for the NVDA screen reader
Added a basic driver for the NVDA screen reader.
Still needs a bit of work (pitch and inflection values are not adjusted for each voice), but it should get the job done.
Once you've built dectalk.dll as described in the Readme, copy all of the contents from the "build" folder in plat/win32/build (I recommend the "release" branch), to make the driver operational.