Re-Added the NVDA driver
Re-added the NVDA driver. Turns out it wasn't a problem with the synthesizer on my end, I just had a version of DECtalk that conflicted with DECtalk 5.1.
This particular driver was modified from the driver that's included in the DECtalk revival project's repo. The only thing that has been changed was that pitch and inflection for each voice has been readjusted to match those found in p_us_vdf1.h.